Bug 1862295 - Translations UI Integration Integrate Error States
